+++ l10n_bg_claude_terminal/models/__init__.py
@@ -4,6 +4,7 @@
 from . import res_company
 from . import res_config_settings
 from . import res_users
+from . import ir_model
 from . import ai_view_registry
 from . import ai_composite_document
 from . import ai_view_parser